backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pillow
python-dotenv
